[MARGOT] Minor rework on field adaptor and monitors
In case we are going to extract statistical properties over an empty
circular buffer, we no longer trigger an assert, but we got the default
value of the target type.

Moreover, we have introduced the concept of validity. In particular, a
statistical property is valid if it has been extracted from a full monitor.

The field adaptor now uses only valid information to compute the error